file(file): handle remote file size check error in `download_file` resource
